Capturing news reports on websites
Program produced by a Portuguese speaker. If you need to contact us for any questions regarding the code. The function of the softwere is to capture the title and the text of news on the internet, joining with data such as author, website, session and time of the consultation (date and time) for later academic citation.